+++ src/Api/Providers/Interests.php
@@ -3,21 +3,11 @@
 namespace seregazhuk\PinterestBot\Api\Providers;
 
 use seregazhuk\PinterestBot\Helpers\UrlBuilder;
-use seregazhuk\PinterestBot\Api\Traits\Followable;
+use seregazhuk\PinterestBot\Api\Traits\HasRelatedTopics;
 
 class Interests extends Provider
 {
-    use Followable;
-
-    /**
-     * @var array
-     */
-    protected $loginRequiredFor = ['follow', 'unFollow'];
-
-    protected $followUrl   = UrlBuilder::RESOURCE_FOLLOW_INTEREST;
-    protected $unFollowUrl = UrlBuilder::RESOURCE_UNFOLLOW_INTEREST;
-
-    protected $entityIdName = 'interest_id';
+    use HasRelatedTopics;
 
     /**
      * Get list of main categories
@@ -38,5 +28,26 @@ class Interests extends Provider
     public function getInfo($category)
     {
         return $this->execGetRequest(["category" => $category], UrlBuilder::RESOURCE_GET_CATEGORY);
+    }
+
+    /**
+     * Returns a feed of pins
+     * @param string $interest
+     * @param int $limit
+     * @return array|bool
+     */
+    public function getPinsFor($interest, $limit = 0)
+    {
+        $params = [
+            'data' => [
+                'feed'             => $interest,
+                'is_category_feed' => true,
+            ],
+            'url' => UrlBuilder::RESOURCE_GET_CATEGORY_FEED
+        ];
+
+        return $this->getPaginatedResponse(
+          $params, $limit
+        );
     }
 }