+++ tests/Api/InterestsTest.php
@@ -11,8 +11,6 @@ use seregazhuk\PinterestBot\Api\Providers\Interests;
  */
 class InterestsTest extends ProviderTest
 {
-    use FollowResponseHelper;
-
     /**
      * @var Interests
      */
@@ -22,28 +20,6 @@ class InterestsTest extends ProviderTest
      * @var string
      */
     protected $providerClass = Interests::class;
-
-    /** @test */
-    public function it_should_follow_interest()
-    {
-        $interestId = 1111;
-        $this->setFollowSuccessResponse($interestId, UrlBuilder::RESOURCE_FOLLOW_INTEREST);
-        $this->assertTrue($this->provider->follow($interestId));
-
-        $this->setFollowErrorResponse($interestId, UrlBuilder::RESOURCE_FOLLOW_INTEREST);
-        $this->assertFalse($this->provider->follow($interestId));
-    }
-
-    /** @test */
-    public function it_should_unfollow_interest()
-    {
-        $interestId = 1111;
-        $this->setFollowSuccessResponse($interestId, UrlBuilder::RESOURCE_UNFOLLOW_INTEREST);
-        $this->assertTrue($this->provider->unFollow($interestId));
-
-        $this->setFollowErrorResponse($interestId, UrlBuilder::RESOURCE_UNFOLLOW_INTEREST);
-        $this->assertFalse($this->provider->unFollow($interestId));
-    }
 
     /** @test */
     public function it_should_return_main_categories()
@@ -67,5 +43,16 @@ class InterestsTest extends ProviderTest
         $this->setResponseExpectation($response);
 
         $this->assertEquals($info, $this->provider->getInfo(1));
+    }
+
+    /** @test */
+    public function it_should_return_generator_for_pins()
+    {
+        $response = $this->createPaginatedResponse();
+
+        $this->setResponseExpectation($response);
+        $this->setResourceResponseData([]);
+
+        $this->assertCount(2, iterator_to_array($this->provider->getPinsFor('test')));
     }
 }